More general component for displaying one calendar page
The PageGroundFloor component was adapted to serve as a more general 
component that can accept any number of options and display them 
correctly. With this, the API changed, so all the texts and labels are 
now set inside the Month component. There is a new CalendarPageOption 
component that makes adding options easiy.
This also means, we no longer need two very similar components, reducing 
the code duplication.