Improved styling
Pages for one month are now displayed side by side - not yet responsive 
though.
You can add custom classNames to the CalendarPage component and apply 
your own styling. This is how the ground and first floor now have 
different colors. For easing the work with classNames the classnames 
package was installed.
Also the print is now in colours.