More readable ground floor options key names
I adapted the key names for the options in the ground floor as they were 
still legacy from a time when I though to just translate the words, 
without any structure to them. Not they are a bit more streamlined. Also 
some grammar was fixed as well as a description for the creative room 
was added.