Controls for changing the year
Added basic controls to easily change the year the clendar displays. For 
that I had to add the Controls component, add an App component to keep 
it all together and make some changes to the CalendaPage component. The 
latter was not responding to changes in the year. This was because the 
date for each page got set when the component was created and after that 
never updated.