Making things pretty with fonts
I wanted to use a specific font rather than some random default one. I 
chose a font with serifes in hope that it improves the readability once 
the calendar is printed. I used two fonts. The default calendar font 
does not display written out times very well. So I used a more standard 
one for the descriptions of the options.
For the webinterace I chose a non-serif font.
In order to make things fit again, I had to rework the layout of the 
calendar page. Now everything is fitting neatly again.