Printable background color
With this commit I tried to change the background color into an normal 
image, that is just positioned so that it acts as the background. This 
way I hoped to solve  Edge not printing background colors. It worked, 
though now the individual cells don't have a background color as well. 
Annoying. Firefox is still not really working.
I think I'll try using svg's instead, will make a new branch for that.