Removed backdrop date due to renderinge issues
The date that was used as a kind of backdrop for the written out month 
name and year didn't render correctly. (In Chrome anyways, not that 
other browsers would do better.) It was blurry and positioned wrong on 
all pages but the first. (Reminds me of the good old times where every 
browser needed their own style sheets to make anything work.)
In order to make things look okay again, I moved the written out date a 
bit to the left.