Disabled all styles and found a way to print full page in mayor browsers
I finally found a way to always scale the Page div to the full size of 
the page. I thought absolute values to the size of an A4 page would do 
the trick. Works in Chrome, almost works in Edge, but in FF things are 
all over the place.
The solution now is to not rely on the browsers intepretation of 
absolute values, but on dynamic values (vw, vh) instead.
With this base I hope to find a styling that will work across all 
browsers. Oh, did I mention that i just have red, rectangles? Yeah, I 
stripped everything but the necessary, just to be sure to avoid strange 
HTML edge cases breaking my layout in other ways.