Moved over to using relative sizes (rem)
In this commit I changed the sizes of almost everything to rem values.

Why did I do that?
The problem I encountered using absolute values (mm) for all the page 
divs was that even when set to the correct dimensions of 210mm x 297mm 
they still didn't quite fit to the A4 page in the print preview in every 
browser - looking at you FireFox. I got around this problem by using 
relative values instead. With 100vw x 100vh the page div spans the whole 
page and only the whole page - even in FireFox.
But how do I correctly size the content of the page without using 
absolute values?
Relative values instead of absolute ones! In screen mode I can set the 
font size to 1mm and with all values moved from mm to rem nothing will 
change in size.
But now when I print, I can set the font size to 100 divided by 297, 
that is the percentual height of a milimeter comapred to the length of 
an A4 page. This way I can calculate the size in which the browser 
renders a milimeter when printing and use this value rather than an 
absolute value of 1mm. Using relative values for all the content things 
should look the same in every browser.