Renamed / moved files for better structure
I decided to not go with the year, month, calendarPage naming scheme and
use the terminology for a real calendar instead. So Calendar is what was
Year, because a calendar usually contains a full year. CalendarPage
became just Page as it is one page in the calendar and Page became
Sheet, as it really is just modelling the paper on which the page is
printed. Also, all calendar specific components are now in their own
folder.