Files Split & Styles simplified
The bulky Page.jsx is now split into multiple files, each for one 
component. While doing so I realized that JSS is broken when using 
dynamic rules and also creating a ton of rules which are not necessary 
in my case. So all the calculation and modelling of the page layout is 
gone and I rely on a browser based approach where the table is now 100% 
in height and due to the fixed table layout the rows are all teh same 
height. To accommodate for the different number of days in different 
months I added placeholder rows to the end of each month so that all 
month have 31 + header rows.