Printable cell backgrounds for Edge and FF
Edge and FireFox do not print backgrounds applied by CSS. As a 
workaround I used images and a bit of absolute positioning to make it 
appear as if the images were the background.
Now I took this concept to the cells as well. This way we have our 
weekend highlighting in FireFox and in Edge the table no longer is 
transparent. Works quite nice as well. Edge now looks as same a Chrome - 
better than FireFox!
I installed a module to do the browser detection / switching.