[WIP] No more Placeholders
I again had rendering issues with the placeholders in Firefox. So here, I want
to take them out, and only render Day components. This way, all table rows
should be the same and thus take up the same space.

So far I have not adapted the Day component to also act as a placeholder.