Refactored Day component
Here, I refactored the Day component. It is now separated into two components:
The Day component for handling the business logic and the Row component for
displaying.

Doing so, I also implemented the logic to check whether a day is a placeholder
or not and act accordingly.

I also threw out the background handling for Firefox. It appears to work now
without using images as background for the table cells. I think this is much
cleaner and easier to understand.
Support for IE is dropped with that.